+++ crates/maelstrom-macro/src/config.rs
@@ -25,6 +25,8 @@ struct ConfigStructField {
     flatten: bool,
     #[darling(default)]
     var_arg: bool,
+    #[darling(default)]
+    list: bool,
     short: Option<char>,
     alias: Option<String>,
     value_name: Option<String>,
@@ -197,6 +199,26 @@ impl ConfigStructField {
         })
     }
 
+    fn gen_builder_list_call(&self) -> Result<Expr> {
+        let name = self.ident().unraw().to_string();
+        let short = self.short();
+        let alias = self.alias();
+        let value_name = self.value_name()?;
+        let default = self.default_as_string_option();
+        let doc = self.doc_comment()?;
+        Ok(parse_quote! {
+            let builder = ::maelstrom_util::config::CommandBuilder::list(
+                builder,
+                #name,
+                #short,
+                #alias,
+                #value_name,
+                #default,
+                #doc,
+            )
+        })
+    }
+
     fn gen_flatten_add_command_line_options_call(&self) -> Expr {
         let field_type = &self.ty;
         parse_quote! {
@@ -290,6 +312,8 @@ impl ConfigInput {
                     field.gen_builder_option_value_call()
                 } else if field.var_arg {
                     field.gen_builder_var_arg_call()
+                } else if field.list {
+                    field.gen_builder_list_call()
                 } else {
                     field.gen_builder_value_call()
                 });
@@ -319,7 +343,7 @@ impl ConfigInput {
                 field.gen_config_bag_get_flag_call()
             } else if field.option {
                 field.gen_config_bag_get_option_call()
-            } else if field.var_arg {
+            } else if field.list || field.var_arg {
                 field.gen_config_bag_get_list_call()
             } else {
                 field.gen_config_bag_get_call()