+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `extra-test-binary-args` config value added to `maelstrom-go-test`
 - `extra-test-binary-args` config value added to `cargo-maelstrom`
 - `extra-pytest-args` config value added to `maelstrom-pytest`
+- `extra-pytest-collect-args` config value added to `maelstrom-pytest`
+- `extra-pytest-test-args` config value added to `maelstrom-pytest`
 
 #### Changed
 - `ui` test-runner configuration option value now matches the same casing as the CLI option